Major views file structure update
