+++ src/views/Contact.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
 //Component
-import Navbar from '../../../components/Navbar';
+import {Navbar} from '../components';
 
-const NotFound = () => {
+const Contact = () => {
     return(
         <React.Fragment>
             <header>
@@ -11,11 +11,11 @@ const NotFound = () => {
             </header>
             <main>
                 <section className="page">
-                    <p className="page__place-holder">404</p>
+                    <p className="page__place-holder">Contact</p>
                 </section>
             </main>
         </React.Fragment>
     );
 }
 
-export default NotFound;
+export default Contact;
