+++ src/views/Error404.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
 //Component
-import Navbar from '../../components/Navbar';
+import {Navbar} from '../components';
 
-const Contact = () => {
+const NotFound = () => {
     return(
         <React.Fragment>
             <header>
@@ -11,11 +11,11 @@ const Contact = () => {
             </header>
             <main>
                 <section className="page">
-                    <p className="page__place-holder">Contact</p>
+                    <p className="page__place-holder">404</p>
                 </section>
             </main>
         </React.Fragment>
     );
 }
 
-export default Contact;
+export default NotFound;
