+++ src/views/Podcast.jsx
@@ -2,17 +2,20 @@ import React from 'react';
 import {Redirect} from 'react-router-dom';
 
 //Component
-import Navbar from '../../components/Navbar';
-import Post from '../../components/Post';
 import {
+    Post,
+    Navbar,
     AudioPlayer,
     AudioPlayerButton
-} from '../../components/AudioPlayer';
+} from '../components';
 
 //Services
-import {API} from '../../services/mockData';
-import {getCurrentPageID,getCurrentPageType} from '../../services/pageInfo';
-import {filterOverall} from '../../services/filters';
+import {API} from '../services/mockData';
+import {
+    getCurrentPageID,
+    getCurrentPageType,
+    filterOverall
+} from '../services';
 
 export const Podcast = () => {
     const podcast = filterOverall(getCurrentPageID(),'podcast',API);
